+++ src/token-indicator.js
@@ -1,6 +1,6 @@
 import { SpriteID} from './sprite-id.js';
 
-
+import * as Helpers from './helpers.js';
 
 /**
  * Used to handle the indicators for direction for tokens
@@ -35,7 +35,6 @@ export class TokenIndicator {
     rotate(deg) {
         if (!this.sprite) { throw 'No sprite to rotate'; return false;}
         this.sprite.angle=deg;
-
     }
 
     /* -------------------------------------------- */
@@ -71,28 +70,104 @@ export class TokenIndicator {
      * This is the default indicator & style. A small triangle
      */
     generateDefaultIndicator() {
-        let i = new PIXI.Graphics();
+
         let indicator_color = colorStringToHex("FF0000");
 
         if (this.token.actor.isPC) {
-            indicator_color = colorStringToHex(getTokenOwner(this.token)[0].color);
+            indicator_color = colorStringToHex(Helpers.getTokenOwner(this.token)[0].color);
         }
-        //indicator_color = colorStringToHex(game.user.color);
+        let triangle = this.generateTriangleIndicator("normal",indicator_color,0x000000);
 
-        i.beginFill(indicator_color, .5).lineStyle(2, 0x000000, 1)
-            .moveTo(this.token.w / 2, this.token.h + 25)
-            .lineTo(this.token.w / 2 - 10, this.token.h + 10)
-            .lineTo(this.token.w / 2 + 10, this.token.h + 10)
-            .lineTo(this.token.w / 2, this.token.h + 25)
-            .closePath()
-            .endFill()
-            .beginFill(0x000000, 0).lineStyle(0, 0x000000, 0)
-            .drawCircle(this.token.w / 2, this.token.w / 2, this.token.w * 2.5)
-            .endFill();
+        return triangle;
+        //indicator_color = colorStringToHex(game.user.color);
+    }
+
+    /**
+     * 
+     * @param {string} size        -- string from ['small','normal','large']
+     * @param {string} fillColor   -- string in hex color code of fill color
+     * @param {string} borderColor -- string in hex color code of border color
+     */
+    generateTriangleIndicator(size="",fillColor="", borderColor="") {
+        let i = new PIXI.Graphics();
+
+        let modHeight = 25;
+        let modWidth  = 10;
+
+        if (size == 'large') {
+            modHeight = 40;
+            modWidth  = 16;
+        }
+
+        i.beginFill(fillColor, .5).lineStyle(2, borderColor, 1)
+        .moveTo(this.token.w / 2, this.token.h + modHeight)
+        .lineTo(this.token.w / 2 - modWidth, this.token.h + modWidth)
+        .lineTo(this.token.w / 2 + modWidth, this.token.h + modWidth)
+        .lineTo(this.token.w / 2, this.token.h + modHeight)
+        .closePath()
+        .endFill()
+        .beginFill(0x000000, 0).lineStyle(0, 0x000000, 0)
+        .drawCircle(this.token.w / 2, this.token.w / 2, this.token.w * 2.5)
+        .endFill();
 
         let texture = canvas.app.renderer.generateTexture(i);
         return new SpriteID(texture, this.token.id);
     }
+
+    generateSpaceIndicator(size="",fillColor="") {
+        let i = new PIXI.Graphics();
+
+
+        i.beginFill(0x000000, .8).lineStyle(2, 0x000000,1)
+        .moveTo(this.token.w/2, 0)
+        .lineTo(this.token.w/2, 0)
+        .closePath()
+        .endFill()
+        .beginFill(fillColor, .8).lineStyle(0, 0x000000, 1)
+        .drawCircle(this.token.w/2, 500, 20)
+        .endFill();
+
+        let texture = canvas.app.renderer.generateTexture(i);
+        return new SpriteID(texture, this.token.id);
+    }
+
+    generateStarIndicator(fillColor=0xe8FF00,borderColor=0x000000) {
+        let i = new PIXI.Graphics();
+        let w = this.token.w;
+        let h = this.token.h;
+        let wc = w/2;
+        let hc = h/2;
+        let arrPoints = [
+            0,0,
+            wc,h,
+            h,wc+100
+        ]
+        
+
+        i.beginFill(fillColor,1).lineStyle(2,borderColor).moveTo(wc,0);
+
+
+        i.drawPolygon([450, -50, // Starting x, y coordinates for the star
+            470, 25, // Star is drawn in a clockwork motion
+            530, 55,
+            485, 95,
+            500, 150,
+            450, 120,
+            400, 150,
+            415, 95,
+            370, 55,
+            430, 25
+          ])
+          .drawCircle(450, 45,60)
+          
+
+        .endFill();
+
+        let texture = canvas.app.renderer.generateTexture(i);
+        return new SpriteID(texture, this.token.id);
+        
+    }
+
 
     /* -------------------------------------------- */
 
@@ -100,11 +175,21 @@ export class TokenIndicator {
      * Create the indicator using the instance's indicator sprite
      * If one hasn't been specified/set, use the default
      */
-    create() {
-        if (!this.sprite) {
+    create(sprite="") {
+        if (!this.sprite && sprite == "") {
             this.sprite = this.generateDefaultIndicator();
+            
+            // this.sprite = this.generateSpaceIndicator('',0x000000);
+            // this.sprite = this.generateStarIndicator();
+        } else if (sprite != "") {
+            if (sprite == "large-triangle") {
+                this.sprite = this.generateTriangleIndicator('large',0xEAFF00,0x000000);
+            } else {
+                this.sprite = this.generateDefaultIndicator();
+            }
         }
 
+        this.sprite.zIndex = -1;
         this.sprite.position.x = this.token.w/2;
         this.sprite.position.y = this.token.w/2;
         this.sprite.anchor.set(.5);
@@ -117,24 +202,3 @@ export class TokenIndicator {
 
 }
 
-/*** Utility Stuff, will be hoisted ***/
-function getKeyByValue(object, value) {
-    return Object.keys(object).filter(key => object[key] === value);
-}
-
-function getTokenOwner(token, includeGM=false) {
-    let owners = getKeyByValue(token.actor.data.permission,3);
-    let ret = [];
-    for (let y = 0; y < owners.length; y++) {
-        let u = Users.instance.get(owners[y]);
-        if (includeGM) {
-            ret.push(u);
-            continue;
-        } else {
-            if (!u.isGM) { ret.push(u);}
-        }
-        
-    }
-    return ret;
-
-}