+++ changelog.md
@@ -1,5 +1,22 @@
 # CHANGELOG
 
+
+
+## [1.0.5] 2020-06-18
+
+### ADDED
+- Vertical and Horizontal flip token options (doesn't work with Indicators)
+- Added larger indicator option
+- Backend support for multiple graphics indicators
+
+## [1.0.3] 2020-06-18
+
+### BUGFIXES
+
+## [1.0.2] 2020-06-18
+
+### BUGFIXES
+
 ## [1.0.1] 2020-06-18
 
 ### ADDED